Raise an exception if the driver failed to load
We only try one attempt a loading a driver, so either it succeeds or it
fails badly. Let's throw the appropriate exceptions for that case.

And to do so make name/path optional in the UnsupportedDevice exception
so we can throw the exception where it happens and enrich it in the
caller where we actually have that data.