Revamp the commit approach again
Similar to the previous attempt, we generate a cookie and return it. But
we also take a callback that is passed through to the driver [1] and
called when it is done. This is a narrower approach than the previous
attempt at a commit-complete signal which would also notify other
listeners.

[1] we wrap that one to reset the dirty state but anyway...