Create the Rodent in the main ratbag context
All the drivers will likely use the rodent anyway to access hidraw
devices, so let's automate all this so the drivers get everything
pre-prepared.

For the future use-cases of a driver connecting to something other than
a hidraw device, we can subclass rodent to match whatever connection we
need.

This pushes a DeviceInfo into the Rodent though it's abstracted in the
API.