Massive rework of the architecture
This is one giant commit because it's too hard to separate all this out.
The main goal here is to make the drivers more independent from the
context and have the context be as simple as possible, without having to
care for the driver implementation.

So the way it works now is:
- drivers use the @ratbag_driver("drivername") decorator on the main
  class
- drivers are instantiated through a class method on that driver class,
  this allows for easier addition of other APIs later
- drivers simply get the "start" signal from the context, and the rest
  is up to them
- the context loads all data files on startup, extracts the driver
  names, loads those modules and instantiates all drivers with the list
  of known devices

The context has no other knowledge of anything, which devices are added
and how is completely up to the driver.

Since most our drivers listen for hidraw devices, a convenience class is
provided that lists/notifies hidraw devices. That class has hooks for
the emulator so we can plug fake devices in.

Note that this currently breaks the recorders, not sure what to do about
those yet.