emulator: fix the emulator for set/get feature with identical data
On the Roccat KONE, the driver may use the same set of bytes for the
GetFeature request as for the SetFeature request. With the previous
implementation we'd fail because the latter would add to the former's
data fields and we run out of values on replay.

Fix this by separating ioctls by name (and cleaning up a bunch of other
things).