Sync the data files from libratbag to here
As of libratbag commit 1e3a85433af41955424f237af1975243edeabefa

Due to Python packaging constraints (or me not knowing how to work
around them) the files need to be inside the ratbag package directory.

The only exception is the logo.svg which we'll keep inside the data
directory until we know what to do with it.