parser: if we're de-parsing an array of zeros, don't spam the log
The roccat driver uses the Parser to initialize some instances directly
(rather than using a intermediate object like the hidpp20 driver). This
requires that it initializes with an array of zeroes first to ensure all
attributes in the Spec exist.

For that case, let's disable the logger automatically since we don't
need the full output of which zero is assigned to where, the Spec tells
us the format anyway.