parser: use a temporary dict for values and construct a proper result class
Instead of calling setattr as we parse, store all the values in a dict
and then, once complete, store the attributes on the result object. This
means we won't have partially set objects if a parsing failure happens.

And because we now have all names and values easily available at the end
of the parsing we can create a proper class from this approach and get
to benefit from attr's default handling of things like __str__ and
similar.