Add a udev rule + hwdb file generator for uaccess
This script enables us to generate a hwdb file of all devices in our
data base. The hwdb file simply sets RATBAG_DEVICE=1, the accompanying
udev file applies the uaccess tag which means the device is readable by
a logged-in user. No more need to run ratbag* as root. With all the
other security issues that implies of course.